refactor(configuration): Change folder structure. Each class has a header
