+++ src/Configuration/Directive/Block.hpp
@@ -1,18 +1,12 @@
 #pragma once
 
 #include <map>
-#include <vector>
 #include <utility>
-
-#include "misc/Maybe.hpp"
 
 #include "Configuration/Directive.hpp"
 
 namespace directive
 {
-  class ServerBlock;
-  class LocationBlock;
-
   /**
    * @brief Directives is a map of directive names to their corresponding
    * Directive objects. It is possible to have multiple entry for the same
@@ -22,18 +16,6 @@ namespace directive
   typedef std::multimap<Directive::Type, Directive*> Directives;
 
   typedef std::pair<Directives::const_iterator, Directives::const_iterator> DirectivesRange;
-
-  /**
-   * @brief Servers points to a range of Directives that are of type
-   * DirectiveServer.
-  */
-  typedef DirectivesRange Servers;
-
-  /**
-   * @brief Locations points to a range of Directives that are of type
-   * DirectiveLocation.
-  */
-  typedef DirectivesRange Locations;
 
   class DirectiveBlock : public Directive
   {
