+++ src/Configuration/Directive/Block/Location.cpp
@@ -1,6 +1,8 @@
-#include "Blocks.hpp"
+#include "Location.hpp"
 
 #include "misc/Maybe.hpp"
+#include "Configuration/Directive.hpp"
+#include "Configuration/Directive/Block.hpp"
 
 namespace directive
 {
