+++ src/Configuration/Directive/Block/Location.hpp
@@ -1,0 +1,30 @@
+#pragma once
+
+#include "misc/Maybe.hpp"
+#include "Configuration/Directive/Block.hpp"
+
+namespace directive
+{
+  /**
+   * @brief Locations points to a range of Directives that are of type
+   * DirectiveLocation.
+  */
+  typedef DirectivesRange Locations;
+
+  class LocationBlock : public DirectiveBlock
+  {
+    public:
+      LocationBlock();
+      LocationBlock(const Context& context);
+      LocationBlock(const LocationBlock& other);
+      LocationBlock& operator=(const LocationBlock& other);
+      virtual ~LocationBlock();
+
+      virtual Type          type() const;
+
+      Maybe<Locations>      locations();
+
+    private:
+      Maybe<Locations> locations_;
+  };
+} // namespace configuration
