+++ src/Configuration/Directive/Simple/Listen.cpp
@@ -1,6 +1,9 @@
-#include "Simples.hpp"
+#include "Listen.hpp"
 
 #include <vector>
+
+#include "Configuration/Directive.hpp"
+#include "Configuration/Socket.hpp"
 
 namespace directive
 {
