+++ src/Configuration/Directive/Simple/MimeTypes.cpp
@@ -1,10 +1,11 @@
-#include "Simples.hpp"
+#include "MimeTypes.hpp"
 
-#include <map>
 #include <string>
+#include <map>
 
 #include "misc/Maybe.hpp"
 #include "misc/Nothing.hpp"
+#include "Configuration/Directive.hpp"
 
 namespace directive
 {
